Add C# writer
Note: this does not actually compile due to string limits. The idea was to bake it all in as code-level strings, but dotnet build process seems to allow maximum of 8MB of those strings. Seems silly. See later if there is a workaround.

The library this file was meant for will use json resource approach for now & hopes that dotnet reasonably bakes resources in.